Move string parsing code to wayland_scanner::token
